requirements.txt
apache-airflow-providers-google>=10.0.0
pyspark>=3.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
pandas>=1.3.0
